+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- ...
+- Overlay: Detailed speed readout with percentage of top speed
 
 ## [0.1.6] - 2024-08-11